Release v2.0.0

- Birth of air-quota!

Improvements
- Implement `AmadeusQuotaChecker` class to extract number of remaining tickets and EMDs
  from Amadeus direct stock offices via Amadeus Web Services (Header SOAP4.0).
- Implement `SirenaQuotaChecker` class to make quota check flow for Amadeus and Sirena the same.
- Redesign quota check flow with new classes.

Bugfixes
- Fix verification of recipient email. AWS SES check is case-sensitive.
- Add forgotten account parameter `alert` into conf sample.
- Fix db test. It didn't work during the first run of tests.
- Fix alert check. Now it works correctly in case number of remaining tickets of the db entry is NULL.

+++ tests/test_db.py
@@ -12,7 +12,8 @@ class TestDB():
 
     def test_init_db(self):
         try:
-            os.remove(self.db_name)
+            if os.path.isfile(self.db_name):
+                os.remove(self.db_name)
         except OSError:
             pytest.fail('Unexpected error trying to delete {}'.format(self.db_name), pytrace=True)
         conn = sqlite3.connect(self.db_name)
